Add files via upload
Please, in order to improve the chances for a successful build, delete all files that are not code(.cs) and in VS from File>New>Project From Existing Code select the folder structure and use it. Please don't delete Skype4COM.dll from release and don't move the chess figures folder, because the code uses System.IO.Directory.GetCurrentDirectory() to refer to the folder the executable is being build. In order to register the library and be able to build the project, please, right click on project's name in Solution Explorer > Properties > Reference Paths add the bin/release folder. It should be a relative path. Then go to Solution Explorer > References right click Add Reference.. and search in the COM part for Skype4COM.dll .  When fond, mark it to add a reference to the library and build the project. Then it should work. And sending the folder release to somebody else's computer should be enough to start the application too. (Don't forget to give it access from inside Skype..)